Fix wrong address data type size
Should have been int instead of short as unsigned short is not large
enough for unsigned 16 bit addressing.

+++ src/main/java/com/github/reisnera/gameboylfb/GameBoyAppLauncher.java
@@ -41,9 +41,9 @@ public class GameBoyAppLauncher {
 		}
 
 		// testing...
-		System.out.println(String.format("%x", mem.readByte((short)0)));
+		System.out.println(String.format("%x", mem.readByte(0)));
 		mem.disableDmgRom();
-		System.out.println(String.format("%x", mem.readByte((short)0)));
+		System.out.println(String.format("%x", mem.readByte(0)));
 	}
 
 	private static void configureLogging() {
